requirements.txt
python-telegram-bot
requests
pyyaml
selenium
marshmallow
argparse
